Introduce data store abstraction layer
1. Add dot env DATASTORE that controls which implementation to load
   from './ds/[DATASTORE].js'.

2. Load sequence: Controller > ds/datastore.js > ds/[DATASTORE].js

3. Add dummy and mongo data stores.

Dummy data store give static response to allow quickly verifying
that everything is working as expected.

Mongo data store preserves the original storage made to MongoDB.
