+++ .env.example
@@ -1,2 +1,4 @@
 TELEGRAM_BOT_TOKEN='TELEGRAM_BOT_TOKEN'
 DISABLE_LOGGING=true
+# Choose the data store implementation from the ds directory.
+DATASTORE=mongo
