+++ .travis.yml
@@ -1,6 +1,7 @@
 language: node_js
 node_js:
   - node
+  - 5
 after_success:
   - 'npm run coveralls'
   - codecov
