+++ test/startTest.js
@@ -1,6 +1,16 @@
 require('codecov');
 var log = require('log-to-file-and-console-node');
 log.removeConsole();
+
+require('dotenv').load();
+
+// Since controller/testMessage is coupled with mongoose, only the 'mongo'
+// data store will pass the test.
+process.env.DATASTORE = 'mongo';
+
+// datastore, must be executed first to stub mongoose.connect(...)
+require('./ds/testMongo');
+
 // express
 // TODO: add test case for app.js
 // route
